refactor: Collapse nested test layout to sibling tests/ for orbit-store… [ORB-00243]
Collapse nested test layout to sibling tests/ for orbit-store and orbit-tools

Planned-By: grok
