+++ crates/orbit-store/src/sqlite/invocation_store.rs
@@ -9,3 +9,6 @@ pub use types::{
     ActivityInvocationMetrics, AgentInvocationMetrics, InvocationInsertParams, InvocationQuery,
     InvocationRecord, InvocationToolCallRecord, TaskInvocationMetrics, ToolInvocationMetrics,
 };
+
+#[cfg(test)]
+mod tests;
