+++ crates/orbit-store/src/sqlite/invocation_store/tests/records.rs
@@ -1,7 +1,10 @@
-// Migrated from sqlite/invocation_store/records.rs per ORB-00231
+// Migrated from sqlite/invocation_store/records/tests/records.rs (anti-pattern child of source)
+// to sibling under `invocation_store/tests/` per ORB-00243 and
+// docs/design-patterns/test_layout.md. (Consolidated; single original file.)
+
 use orbit_common::types::{InvocationTrace, RoleSlot, TokenUsage, ToolCallTrace};
 
-use super::super::*;
+use super::super::records::*;
 
 #[test]
 fn invocation_records_persist_planning_duel_slot() {
