+++ crates/orbit-tools/src/builtin/github/pr_comment_reply.rs
@@ -27,7 +27,9 @@ pub(super) fn build_exec_request(
     Ok(super::gh_exec_request(args, None, TIMEOUT_DEFAULT_MS))
 }
 
-fn parse_reply_response(stdout: &str) -> Result<Value, OrbitError> {
+// pub(super) visibility widened from private so that github::tests::pr_comment_reply (sibling test after nested collapse)
+// can invoke the helper. See ORB-00243 and docs/design-patterns/test_layout.md.
+pub(super) fn parse_reply_response(stdout: &str) -> Result<Value, OrbitError> {
     let id = super::parse_gh_api_id(stdout, "gh api (pr comment reply)")?;
     Ok(json!({
         "id": id,
@@ -54,5 +56,3 @@ super::gh_tool! {
     }
 }
 
-#[cfg(test)]
-mod tests;
