--- crates/orbit-tools/src/builtin/github/pr_review/tests/mod.rs
@@ -1,3 +1,0 @@
-#![allow(missing_docs)]
-
-mod response_parsing;
