+++ crates/orbit-tools/src/builtin/github/tests/mod.rs
@@ -1,0 +1,5 @@
+#![allow(missing_docs)]
+
+mod pr_comment_reply;
+mod pr_review;
+mod pr_review_comment;
