+++ crates/orbit-tools/src/builtin/orbit/friction/add.rs
@@ -55,5 +55,3 @@ impl Tool for OrbitFrictionAddTool {
     }
 }
 
-#[cfg(test)]
-mod tests;
