--- crates/orbit-tools/src/builtin/orbit/friction/add/tests/mod.rs
@@ -1,3 +1,0 @@
-#![allow(missing_docs)]
-
-mod tags;
