+++ crates/orbit-tools/src/builtin/orbit/friction/tests/add.rs
@@ -1,8 +1,13 @@
 //! Tags parameter description tests for friction.add.
+//
+// Migrated from nested `friction/add/tests/` (anti-pattern child of source)
+// to sibling layout under `friction/tests/` per ORB-00243 and
+// docs/design-patterns/test_layout.md.
 
 use orbit_common::friction::{DEFAULT_FRICTION_TAGS, friction_tags_literal};
 
-use super::super::*;
+use super::super::add::*;
+use crate::Tool;
 
 #[test]
 fn tags_parameter_description_lists_default_taxonomy() {
