+++ crates/orbit-tools/src/builtin/orbit/friction/tests/mod.rs
@@ -1,3 +1,3 @@
 #![allow(missing_docs)]
 
-mod diagnostics;
+mod add;
