+++ crates/orbit-tools/src/builtin/orbit/graph_history.rs
@@ -49,5 +49,3 @@ impl Tool for OrbitGraphHistoryTool {
     }
 }
 
-#[cfg(test)]
-mod tests;
