+++ crates/orbit-tools/src/builtin/orbit/knowledge/tests/pack.rs
@@ -1,8 +1,12 @@
 //! Tests for refresh diagnostics in pack output.
+//
+// Migrated from nested `knowledge/pack/tests/` (anti-pattern child of source)
+// to sibling under `knowledge/tests/` per ORB-00243 and
+// docs/design-patterns/test_layout.md.
 
 use orbit_knowledge::KnowledgePackResult;
 
-use super::super::add_refresh_diagnostics;
+use super::super::pack::add_refresh_diagnostics;
 
 #[test]
 fn refresh_diagnostics_only_report_actual_auto_refresh_skips() {
