--- crates/orbit-tools/src/builtin/orbit/search/tests/mod.rs
@@ -1,3 +1,0 @@
-#![allow(missing_docs)]
-
-mod schema;
