+++ crates/orbit-tools/src/builtin/orbit/task/tests/mod.rs
@@ -1,0 +1,4 @@
+#![allow(missing_docs)]
+
+mod artifact_put;
+mod update;
