+++ crates/orbit-tools/src/builtin/orbit/task/update.rs
@@ -161,5 +161,3 @@ impl Tool for OrbitTaskUpdateTool {
     }
 }
 
-#[cfg(test)]
-mod tests;
