--- crates/orbit-tools/src/builtin/orbit/task/update/tests/mod.rs
@@ -1,4 +1,0 @@
-#![allow(missing_docs)]
-
-mod context_files;
-mod source_task_id;
