+++ crates/orbit-tools/src/builtin/orbit/tests/graph_history.rs
@@ -1,6 +1,14 @@
 //! Schema tests for the deprecated graph history tool.
+//
+// Migrated from nested `graph_history/tests/` (anti-pattern child of source)
+// to sibling layout under `orbit/tests/` per ORB-00243 and
+// docs/design-patterns/test_layout.md.
+//
+// The structural change ensures tests only reach `pub` / `pub(crate)` / `pub(super)`
+// items of `graph_history`.
 
-use super::super::*;
+use super::super::graph_history::*;
+use crate::Tool;
 
 #[test]
 fn schema_lists_required_selector_without_task_id_pattern() {
