+++ crates/orbit-tools/src/builtin/orbit/tests/mod.rs
@@ -1,3 +1,5 @@
 #![allow(missing_docs)]
 
+mod graph_history;
 mod identity;
+mod search;
