+++ crates/orbit-tools/src/builtin/orbit/tests/search.rs
@@ -1,6 +1,11 @@
 //! Schema tests for orbit.search tool.
+//
+// Migrated from nested `search/tests/` (anti-pattern child of source)
+// to sibling layout under `orbit/tests/` per ORB-00243 and
+// docs/design-patterns/test_layout.md.
 
-use super::super::*;
+use super::super::search::*;
+use crate::Tool;
 
 #[test]
 fn search_schema_uses_hybrid_and_semantic_task_id_params() {
